feat : zustandの導入

+++ src/components/PostList.tsx
@@ -1,20 +1,16 @@
-
-import { COLLECTION, PostData, PostListItem } from "@/types/types";
-import { Agent } from '@atproto/api';
-import { useEffect, useState } from "react";
 import PostTextWithBold from "@/components/PostTextWithBold";
+import { formatDateToLocale } from "@/logic/LocaledDatetime";
+import { useAtpAgentStore } from "@/state/AtpAgent";
+import { useLocaleStore } from "@/state/Locale";
+import { COLLECTION, PostData, PostListItem } from "@/types/types";
+import Image from 'next/image';
+import { useEffect, useState } from "react";
 
 type DeleteListProps = {
-    agent: Agent;
-    locale: any,
-    did: string,
     handleEdit: (input: PostListItem) => void;
 };
 
 export const DeleteList: React.FC<DeleteListProps> = ({
-    agent,
-    locale,
-    did,
     handleEdit
 }) => {
     const [cursor, setCursor] = useState("");
@@ -23,18 +19,26 @@ export const DeleteList: React.FC<DeleteListProps> = ({
     const [isDeleteing, setIsDeleting] = useState<boolean>(false)
     const [selectedItem, setSelectedItem] = useState<PostListItem | null>(null);
     const [duplicate, setDuplicate] = useState<boolean>(false);
+    const did = useAtpAgentStore((state) => state.did);
+    const agent = useAtpAgentStore((state) => state.agent);
+    const locale = useLocaleStore((state) => state.localeData);
 
     const getPosts = async (did: string, cursor: string) => {
 
         if (duplicate) {
             return;
         }
+        if (!agent) {
+            console.error("未ログインです")
+            return
+        }
+
         setDuplicate(true); // 重複実行を防ぐ
 
         setIsLoading(true)
         setDeleteList([])
         //console.log('call');
-        let deleteList: any[] = []; // 初期化
+        const deleteList: any[] = []; // 初期化
         try {
             const param = {
                 repo: did,
@@ -49,7 +53,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
             setCursor(bookMark.data.cursor || '');
 
             // records を処理して deleteList を更新
-            for (let obj of bookMark.data.records) {
+            for (const obj of bookMark.data.records) {
                 const value = obj.value as PostData;
                 deleteList.push({
                     blurATUri: obj.uri,
@@ -62,7 +66,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
             deleteList.sort((a, b) => {
                 return new Date(b.blur.createdAt).getTime() - new Date(a.blur.createdAt).getTime();
             });
-            
+
             // setDeleteList を呼び出して UI を更新
             setDeleteList(deleteList);
             setIsLoading(false);
@@ -100,6 +104,11 @@ export const DeleteList: React.FC<DeleteListProps> = ({
     };
 
     const deleteRecord = async (aturi: string) => {
+        if (!agent) {
+            console.error("未ログインです")
+            return
+        }
+
         const param = convertAtUrlToObject(aturi)
         await agent.com.atproto.repo.deleteRecord(param)
 
@@ -143,26 +152,11 @@ export const DeleteList: React.FC<DeleteListProps> = ({
         return ''
     };
 
-
-    const formatDateToLocale = (dateString: string) => {
-        const date = new Date(dateString);
-        const userLocale = navigator.language; // ブラウザのロケールを取得
-
-        return new Intl.DateTimeFormat(userLocale, {
-            year: "numeric",
-            month: "numeric",
-            day: "numeric",
-            hour: "numeric",
-            minute: "numeric",
-            second: "numeric",
-            hour12: false, // 24時間表示
-        }).format(date);
-    };
-
     let useEffectDuplidate = false
 
     useEffect(() => {
         if (useEffectDuplidate) return
+        // eslint-disable-next-line react-hooks/exhaustive-deps
         useEffectDuplidate = true
         // fetchBookmarks を呼び出す
         console.log('useEffect')
@@ -172,7 +166,8 @@ export const DeleteList: React.FC<DeleteListProps> = ({
         // 実行後、duplicate を再度 false に設定
         setDuplicate(false);
 
-    }, [did]); // `did` または `cursor` が変更された場合に再実行
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
 
     return (
         <>
@@ -191,13 +186,13 @@ export const DeleteList: React.FC<DeleteListProps> = ({
                             className="py-3 px-2 mb-2 mx-2 bg-white rounded-md border border-gray-400 w-full "
                         >
                             <div>
-                                <PostTextWithBold postText={item.blur.text} isValidateBrackets={true}/>
+                                <PostTextWithBold postText={item.blur.text} isValidateBrackets={true} />
                             </div>
                             <div className="flex justify-between gap-2 mt-2">
                                 <div className="text-sm text-gray-400">{formatDateToLocale(item.blur.createdAt)}</div>
                                 <div className="flex gap-2">
                                     <div className="text-sm text-red-500 mx-3" onClick={() => handleSelectItem(item)}>
-                                        <img
+                                        <Image
                                             src="https://backet.skyblur.uk/trash2.svg" // public フォルダ内のファイルは / からの相対パスで指定
                                             alt="Trash Icon"
                                             width={20} // 必要に応じて幅を指定
@@ -205,7 +200,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
                                         />
                                     </div>
                                     <div onClick={() => handleEdit(item)} className="text-sm text-gray-500 mx-3">
-                                        <img
+                                        <Image
                                             src="https://backet.skyblur.uk/edit.svg" // public フォルダ内のファイルは / からの相対パスで指定
                                             alt="Export Icon"
                                             width={20} // 必要に応じて幅を指定
@@ -213,7 +208,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
                                         />
                                     </div>
                                     <a className="text-sm text-gray-500 mx-3" href={item.blurURL} target="_blank">
-                                        <img
+                                        <Image
                                             src="https://backet.skyblur.uk/right-arrow.svg" // public フォルダ内のファイルは / からの相対パスで指定
                                             alt="Export Icon"
                                             width={20} // 必要に応じて幅を指定
@@ -221,7 +216,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
                                         />
                                     </a>
                                     <a className="text-sm text-gray-600 ml-2 mr-2" href={item.postURL} target="_blank">
-                                    <svg className="h-5 w-5" width="20" height="20" viewBox="0 0 1452 1452" xmlns="http://www.w3.org/2000/svg"><path d="M725.669,684.169c85.954,-174.908 196.522,-329.297 331.704,-463.171c45.917,-43.253 98.131,-74.732 156.638,-94.443c80.779,-23.002 127.157,10.154 139.131,99.467c-2.122,144.025 -12.566,287.365 -31.327,430.015c-29.111,113.446 -96.987,180.762 -203.629,201.947c-36.024,5.837 -72.266,8.516 -108.726,8.038c49.745,11.389 95.815,32.154 138.21,62.292c77.217,64.765 90.425,142.799 39.62,234.097c-37.567,57.717 -83.945,104.938 -139.131,141.664c-82.806,48.116 -154.983,33.716 -216.529,-43.202c-28.935,-38.951 -52.278,-81.818 -70.026,-128.603c-12.177,-34.148 -24.156,-68.309 -35.935,-102.481c-11.779,34.172 -23.757,68.333 -35.934,102.481c-17.748,46.785 -41.091,89.652 -70.027,128.603c-61.545,76.918 -133.722,91.318 -216.529,43.202c-55.186,-36.726 -101.564,-83.947 -139.131,-141.664c-50.804,-91.298 -37.597,-169.332 39.62,-234.097c42.396,-30.138 88.466,-50.903 138.21,-62.292c-36.46,0.478 -72.702,-2.201 -108.725,-8.038c-106.643,-21.185 -174.519,-88.501 -203.629,-201.947c-18.762,-142.65 -29.205,-285.99 -31.328,-430.015c11.975,-89.313 58.352,-122.469 139.132,-99.467c58.507,19.711 110.72,51.19 156.637,94.443c135.183,133.874 245.751,288.263 331.704,463.171Z" fill="currentColor" /></svg>
+                                        <svg className="h-5 w-5" width="20" height="20" viewBox="0 0 1452 1452" xmlns="http://www.w3.org/2000/svg"><path d="M725.669,684.169c85.954,-174.908 196.522,-329.297 331.704,-463.171c45.917,-43.253 98.131,-74.732 156.638,-94.443c80.779,-23.002 127.157,10.154 139.131,99.467c-2.122,144.025 -12.566,287.365 -31.327,430.015c-29.111,113.446 -96.987,180.762 -203.629,201.947c-36.024,5.837 -72.266,8.516 -108.726,8.038c49.745,11.389 95.815,32.154 138.21,62.292c77.217,64.765 90.425,142.799 39.62,234.097c-37.567,57.717 -83.945,104.938 -139.131,141.664c-82.806,48.116 -154.983,33.716 -216.529,-43.202c-28.935,-38.951 -52.278,-81.818 -70.026,-128.603c-12.177,-34.148 -24.156,-68.309 -35.935,-102.481c-11.779,34.172 -23.757,68.333 -35.934,102.481c-17.748,46.785 -41.091,89.652 -70.027,128.603c-61.545,76.918 -133.722,91.318 -216.529,43.202c-55.186,-36.726 -101.564,-83.947 -139.131,-141.664c-50.804,-91.298 -37.597,-169.332 39.62,-234.097c42.396,-30.138 88.466,-50.903 138.21,-62.292c-36.46,0.478 -72.702,-2.201 -108.725,-8.038c-106.643,-21.185 -174.519,-88.501 -203.629,-201.947c-18.762,-142.65 -29.205,-285.99 -31.328,-430.015c11.975,-89.313 58.352,-122.469 139.132,-99.467c58.507,19.711 110.72,51.19 156.637,94.443c135.183,133.874 245.751,288.263 331.704,463.171Z" fill="currentColor" /></svg>
                                     </a>
                                 </div>
                             </div>
@@ -254,7 +249,7 @@ export const DeleteList: React.FC<DeleteListProps> = ({
                                         {locale.DeleteList_CancelButton}
                                     </button>
                                 </div>
-                                <PostTextWithBold postText={selectedItem.blur.text}isValidateBrackets={true}/>
+                                <PostTextWithBold postText={selectedItem.blur.text} isValidateBrackets={true} />
                             </div>
                         </div>
                     )}
